Wrap the Rust FFI in an idiomatic API
As discussed in #37.

This doesen't yet eliminate the C
layer, because it seems very difficult to allocate and construct in C++
a `Vec<T>` (its layout is unspecified so that its fields can be
reordered by rustc, and it is incompatible with malloced buffers), but
all these horrible `Array<T>` are now hidden in a new `ffi` module.

Nevertheless, this greatly simplifies the API usage, as seen in the
`prologin.rs` test.