rules: client id refactor (#82)
Introduce a new exported function: rules_config().

It enables the rules to specify the number of expected players for this
match. With that information, the server can allocate the player
client_ids to be in natural order: the first player will have id 0,
second 1, etc. This introduces the guarantee that player ids can be used
as vector indices (0-indexed and contiguous), and therefore simplifies
games internal logic.