rules: serverloop: Add timestamp to dumps
The goal is to add a notion of timestamp to the creation of a dump,
that would describe the round, the player and the action if there are
some. It comes with doing dumps much more often.

The use of a map is only a proof of concept to launch some discussion,
as it gives quite a lot of flexibility for what the timestamp can
represent.